# Backend/Tech Dev/swagger.yml
openapi: 3.0.1
# Esta sección contiene información sobre el API
info:
  title: API Turismo UAM
  version: '1.0_20220408'
  description: Especificación de API con Open API para el proyecto de Temas Selectos de Ingeniería de Software.
  contact:
    email: pwngigo@gmail.com
    name: Tech Dev
    
# Esta sección contiene información sobre los lugares
paths:
  
######################
# Lugares Turisticos #
######################
  /v1/lugares:
#getLugares
    get:
      tags: # Los tags permiten agrupar las rutas en el swagger
      - Lugares Turísticos
      summary: HU-1 y HU-3:Recupera la lista completa de lugares turísticos
      description: 'Regresa un objecto que contiene todos los lugares turísticos'
      # El operationId generalmente corresponde al nombre del método en el controller
      operationId: 'getPlaces'
      responses:
        '200':
          description: Consulta exitosa!
          content:
            application/json:
              schema: 
                type: object
                properties:
                  lugares:
                    type: array
                    items:
                        type: object
                        properties:
                          id:
                            type: string
                            example: 12345678
                          nombre:
                            type: string
                            example: Revolucion
                          municipio:
                            type: string
                            example: Centro
                          estado:
                            type: string
                            example: CDMX
                          descripcion:
                            type: string
                            example: Monumento de la revolucion
                          tipo:
                            type: string
                            example: monumento
                          latitud:
                            type: string
                            example: -99.00
                          longitud:
                            type: string
                            example: 12.00
                          horarios:
                            type: string
                            example: De 9 a 5 pm
                          precio:
                            type: number
                            example: 50
                          comentarios:
                            type: array
                            items:
                              type: object
                              properties:
                                usuario:
                                  type: string
                                  example: Rodrigo
                                fecha:
                                  type: string
                                  example: fecha...
                                descripcion:
                                  type: string
                                  example: Yo creo que...
                        
        '404':
            description: No se ha encontrado...
            content:
              application/json:
                schema:
                  "$ref": "#/components/schemas/errorDto"
        '500':
            description: Error interno durante la consulta
            content:
              application/json:
                schema:
                  "$ref": "#/components/schemas/errorDto"

#getLugar
  /v1/lugares/{lugarId}:
    parameters: 
      - "$ref": "#/components/parameters/lugarId"
    get:
      tags: # Los tags permiten agrupar las rutas en el swagger
      - Lugares Turísticos
      summary: HU-2 y HU-6:Recupera la información de un lugar turístico
      description: 'Regresa un objecto que contiene un lugar turístico'
      # El operationId generalmente corresponde al nombre del método en el controller
      operationId: 'getPlace'
      
      responses:
        '200':
          description: Consulta exitosa
          # La estructura del dto está dada por un schema (ver sección components/schemas al final del archivo)
          content:
            application/json:
              schema: 
                "$ref": "#/components/schemas/lugarDto"
        '500':
            description: Error interno durante la consulta
            content:
              application/json:
                schema:
                  "$ref": "#/components/schemas/errorDto"

  # postComentarios
  /v1/lugares/{lugarId}/comentarios:
      parameters: 
      - "$ref": "#/components/parameters/lugarId"
      post:
        tags:
        - Lugares Turísticos
        summary: HU-4:Permite dar de alta un comentario en un lugar en especifico
        description: 'Permite dar de alta un comentario'
        operationId: 'setComentarios'
        requestBody:
          content:
            application/json:
              schema:
                  "$ref": "#/components/schemas/comentarioDto"
        responses:
          '201':
            description: Creación exitosa
            headers:
              Location:
                description: Url del comentario creado
                schema:
                  type: string
            content:
              application/json:
                schema:
                  "$ref": "#/components/schemas/lugarDto"
          '400':
            description: No se pudo crear el comentario
            content:
              application/json:
                schema:
                    "$ref": "#/components/schemas/errorDto"
          '500':
            description: Error interno durante la consulta
            content:
              application/json:
                schema:
                    "$ref": "#/components/schemas/errorDto"  
  
###########
# Hoteles #
###########
  /v1/hotel:
#getHoteles
    get:
      tags: # Los tags permiten agrupar las rutas en el swagger
      - Hoteles
      summary: HU-5:Recupera la lista completa de hoteles
      description: 'Regresa un objecto que contiene todos los hoteles'
      # El operationId generalmente corresponde al nombre del método en el controller
      operationId: 'getHotels'
      responses:
        '200':
          description: Consulta exitosa
          content:
            application/json:
              schema: 
                type: object
                properties:
                  hoteles:
                    type: array
                    items:
                        type: object
                        properties:
                          nombre:
                            type: string
                            example: Hotel Finisterre
                          direccion:
                            type: string
                            example: Calz. de Tlalpan 2043, Parque San Andrés, Coyoacán, 04040 Ciudad de México, CDMX
                          telefono:
                            type: string
                            example: 525556899544
                          sitioWeb:
                            type: string
                            example: http://www.hotelfinisterre.com.mx/
                          descripcion:
                            type: string
                            example: Bulliciosa zona de moda con arquitectura colonial, el Jardín Centenario y el Museo Frida Kahlo.
                          latitud:
                            type: string
                            example: 19.34512356266588
                          longitud:
                            type: string
                            example: -99.14457352718428
                          precio:
                            type: number
                            example: 1500
                          tipo:
                            type: string
                            example: Hotel
                          estacionamiento:
                            type: string
                            example: si
                          piscina:
                            type: string
                            example: si
                          habitacion:
                            type: string
                            example: si
                          lavanderia:
                            type: string
                            example: si
                          wifi:
                            type: string
                            example: si
                          mascotas:
                            type: string
                            example: si
        '500':
            description: Error interno durante la consulta
            content:
              application/json:
                schema:
                  "$ref": "#/components/schemas/errorDto"

                              ########################
                              # Data Transfer Object #
                              ########################

##############
# PARAMETROS #
##############
components:
  parameters: # PARAMETROS
    lugarId: # Lugar ID
      name: lugarId
      in: path
      description: Identificador del lugar generado durante la invocación a POST
      required: true
      schema:
        type: string
        example: 1234
    
    tipoLugar: # tipo lugar ID
      name: tipoLugar
      in: path
      description: Tipo de lugar turistic
      required: true
      schema:
        type: string
        example: Zona arqueologica
        
    hotelId: # Hotel ID
      name: hotelId
      in: path
      description: Identificador del hotel generado durante la invocación a POST
      required: true
      schema:
        type: string
        example: 1234
  
  
############
# ESQUEMAS #
############  
  schemas: 
#lugarDto
    lugarDto:
      type: object
      properties:
        id:
          type: string
          example: 2134457654324
        nombre:
          type: string
          example: Chichén Itzá
        municipio:
          type: string
          example: Tinum
        estado:
          type: string
          example: Yucatán
        descripcion:
          type: string
          example: La ciudad es...
        tipo:
          type: string
          example: Zona arqueologica
        latitud:
          type: string
          example: 20.683032102156055
        longitud:
          type: string
          example: 20.683032102156055
        horarios:
          type: string
          example: Lunes a Viernes de 12:00 a 16:00, Sábado y Domingo de 09:00 a 18:00
        precio:
          type: number
          example: 20
        comentarios:
          type: array
          items:
            type: object
            properties:
              usuario:
                type: string
                example: Rodrigo
              fecha:
                  type: string
                  example: 20/10/2022
              descripcion:
                type: string
                example: blablabla bla
    
    # Inicio de hospedajeDto
    hospedajeDto:
      type: object
      properties:
        nombre:
          type: string
          example: Hotel Finisterre
        direccion:
          type: string
          example: Calz. de Tlalpan 2043, Parque San Andrés, Coyoacán, 04040 Ciudad de México, CDMX
        telefono:
          type: string
          example: 525556899544
        sitioWeb:
          type: string
          example: http://www.hotelfinisterre.com.mx/
        descripcion:
          type: string
          example: Bulliciosa zona de moda con arquitectura colonial, el Jardín Centenario y el Museo Frida Kahlo.
        latitud:
          type: string
          example: 19.34512356266588
        longitud:
          type: string
          example: -99.14457352718428
        precio:
          type: number
          example: 1500
        tipo:
          type: string
          example: Hotel
        estacionamiento:
          type: string
          example: si
        piscina:
          type: string
          example: si
        habitacion:
          type: string
          example: si
        lavanderia:
          type: string
          example: si
        wifi:
          type: string
          example: si
        mascotas:
          type: string
          example: si
#Fin de hospedajeDto

#Inicio de errorDto
    errorDto:
      required:
      - mensaje
      - status
      type: object
      properties:
        timestamp:
          type: string
          description: Estampa de tiempo del error
          example: '2017-01-13T17:09:42.411'
        status:
          maximum: 599
          minimum: 100
          type: integer
          description: Código HTTP del error 
          format: int32
          example: 500
        error:
          type: string
          description: Mensaje enviado por el backend
          example: RuntimeException
        mensaje:
          type: string
          description: Mensaje del error
          example: Se produjo un error al acceder a la BD
      description: Respuesta en caso de error
#Fin deerrorDto

# comentarioDto
    comentarioDto:
      type: object
      properties:
        usuario:
          type: string
          example: Rodrigo
        fecha:
            type: string
            example: 20/10/2022
        descripcion:
          type: string
          example: blablabla bla
#Fin ComentariosDTO

spring:
  datasource:
    url: jdbc:h2:mem:mydb
    username: sa
    password: password
    driverClassName: org.h2.Driver
  jpa:
    spring.jpa.database-platform: org.hibernate.dialect.H2Dialect